backend/pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "LiveKit Voice Agent"
requires-python = ">=3.11"
dependencies = [
    "livekit>=0.12.0",
    "livekit-agents>=0.8.0",
    "livekit-plugins-silero>=0.3.0",
    "livekit-plugins-turn-detector>=0.2.0",
    "livekit-plugins-assemblyai>=0.2.0",
    "livekit-plugins-openai>=0.7.0",
    "livekit-plugins-cartesia>=0.2.0",
    "python-dotenv>=1.0.0",
    "llama-index-core>=0.10.0",
    "llama-index-readers-file>=0.1.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "llama-index-vector-stores-faiss>=0.1.0",
    "faiss-cpu>=1.7.0",
]
[project.scripts]
livekit-agent = "livekit_agent.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/livekit_agent"]
